requirements.txt
aac_metrics==0.4.6
matplotlib==3.7.1
netron==7.3.5
nltk==3.5
nltk==3.8.1
numpy==1.22.3
Pillow==10.0.1
Pillow==10.1.0
rouge==1.0.1
scikit_learn==1.2.2
torch==1.13.1
torchvision==0.14.1
